added fontawesome icons to buttons in index, add, edit pages and add …

+++ delete.php
@@ -1,9 +1,12 @@
 <?php
 	// connect db
 	require('./config.php');
-
-	$statement = $pdo->prepare("DELETE FROM todo WHERE id=".$_GET['id']);
-	$statement->execute();
-
+	// check id first
+	if($_GET['id']) {
+		$statement = $pdo->prepare("DELETE FROM todo WHERE id=".$_GET['id']);
+		$statement->execute();
+	} 
+	
 	header("Location: index.php"); // redirect back to page
+	
 ?>
